Make DrawCardWatcher common
