+++ Mage/src/main/java/mage/abilities/common/DrawNthCardTriggeredAbility.java
@@ -6,15 +6,12 @@ import mage.abilities.effects.Effect;
 import mage.abilities.hint.Hint;
 import mage.abilities.hint.ValueHint;
 import mage.constants.TargetController;
-import mage.constants.WatcherScope;
 import mage.constants.Zone;
 import mage.game.Game;
 import mage.game.events.GameEvent;
 import mage.players.Player;
 import mage.util.CardUtil;
-import mage.watchers.Watcher;
-
-import java.util.*;
+import mage.watchers.common.DrawCardWatcher;
 
 /**
  * @author TheElk801
@@ -37,7 +34,6 @@ public class DrawNthCardTriggeredAbility extends TriggeredAbilityImpl {
 
     public DrawNthCardTriggeredAbility(Zone zone, Effect effect, boolean optional, TargetController targetController, int cardNumber) {
         super(zone, effect, optional);
-        this.addWatcher(new DrawCardWatcher());
         this.targetController = targetController;
         this.cardNumber = cardNumber;
         this.addHint(hint);
@@ -96,36 +92,5 @@ public class DrawNthCardTriggeredAbility extends TriggeredAbilityImpl {
     @Override
     public DrawNthCardTriggeredAbility copy() {
         return new DrawNthCardTriggeredAbility(this);
-    }
-}
-
-class DrawCardWatcher extends Watcher {
-
-    private final Map<UUID, List<UUID>> drawMap = new HashMap<>();
-
-    DrawCardWatcher() {
-        super(WatcherScope.GAME);
-    }
-
-    @Override
-    public void watch(GameEvent event, Game game) {
-        if (event.getType() != GameEvent.EventType.DREW_CARD) {
-            return;
-        }
-        if (!drawMap.containsKey(event.getPlayerId())) {
-            drawMap.putIfAbsent(event.getPlayerId(), new ArrayList<>());
-        }
-        drawMap.get(event.getPlayerId()).add(event.getId());
-    }
-
-    @Override
-    public void reset() {
-        super.reset();
-        drawMap.clear();
-    }
-
-    static boolean checkEvent(UUID playerId, GameEvent event, Game game, int cardNumber) {
-        Map<UUID, List<UUID>> drawMap = game.getState().getWatcher(DrawCardWatcher.class).drawMap;
-        return drawMap.containsKey(playerId) && Objects.equals(drawMap.get(playerId).size(), cardNumber) && event.getId().equals(drawMap.get(playerId).get(cardNumber - 1));
     }
 }
